requirements.txt
web3==7.10.0
python-dotenv==1.1.0
pyyaml==6.0.2
aiofiles==24.1.0